Level class supports Tile Set & Map

+++ src/rect.h
@@ -2,11 +2,11 @@
 #define RECT_H
 
 typedef struct {
-    float x, y;
-    float w, h;
+    int x, y;
+    int w, h;
 } Rect;
 
 int rect_intersection(Rect* a, Rect* b);
-void rect_set(Rect* r, float x, float y, float w, float h);
+void rect_set(Rect* r, int x, int y, int w, int h);
 
 #endif
